requirements.txt
blinker==1.6.2
boto3==1.26.125
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pytz==2023.3
PyMySQL==1.0.3
python-dotenv==1.0.0
requests==2.29.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
Werkzeug==2.3.3
